Fix the date and output if it zero
The date had a bug for when the first line of the valve data was printed
out, it would show today's date.

The output had to be change because the analyzer tool would not bring in
the data correct if the value did not exist from the raw data. When
initial just set everything to 0 but it make the graph look like valve
information was going to zero. This is not true. We now use the previous
line of valve data and insert it if the data is 0.